backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.6.1
python-multipart==0.0.6
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
httpx==0.26.0

# VideoDB and AI dependencies
videodb==0.2.3
openai==1.3.7

# Additional utilities
tqdm==4.66.1
Pillow==10.1.0 